refactor: Easings and Camera to core.
