+++ src/core/easings.ts
